+++ php/modules/timeout_reply_group/backend/Routes.php
@@ -50,18 +50,18 @@ class Routes extends RouterProvider
                 ->middleware(CurrentCorpInfoMiddleware::class)
                 ->middleware(UserRoleMiddleware::class)
                 ->routes(
-                    // 回复规则
-                    Route::get('/reply-rules')->action([ReplyRuleController::class, 'show']),
-                    Route::post('/reply-rules')->action([ReplyRuleController::class, 'save']),
+                // 回复规则
+                    Route::get('/reply-rules')->action([ReplyRuleController::class, 'show'])->defaults(["permission_key" => 'timeout_reply_group.list']),
+                    Route::post('/reply-rules')->action([ReplyRuleController::class, 'save'])->defaults(["permission_key" => 'timeout_reply_group.edit']),
 
                     // 超时规则
-                    Route::get('/rules')->action([RuleController::class, 'list']),
-                    Route::get('/rules/{id:\d+}')->action([RuleController::class, 'show']),
-                    Route::post('/rules')->action([RuleController::class, 'save']),
-                    Route::put('/rules/{id:\d+}')->action([RuleController::class, 'update']),
-                    Route::put('/rules/{id:\d+}/enable')->action([RuleController::class, 'enable']),
-                    Route::put('/rules/{id:\d+}/disable')->action([RuleController::class, 'disable']),
-                    Route::delete('/rules/{id:\d+}')->action([RuleController::class, 'destroy']),
+                    Route::get('/rules')->action([RuleController::class, 'list'])->defaults(["permission_key" => 'timeout_reply_group.list']),
+                    Route::get('/rules/{id:\d+}')->action([RuleController::class, 'show'])->defaults(["permission_key" => 'timeout_reply_group.list']),
+                    Route::post('/rules')->action([RuleController::class, 'save'])->defaults(["permission_key" => 'timeout_reply_group.edit']),
+                    Route::put('/rules/{id:\d+}')->action([RuleController::class, 'update'])->defaults(["permission_key" => 'timeout_reply_group.edit']),
+                    Route::put('/rules/{id:\d+}/enable')->action([RuleController::class, 'enable'])->defaults(["permission_key" => 'timeout_reply_group.edit']),
+                    Route::put('/rules/{id:\d+}/disable')->action([RuleController::class, 'disable'])->defaults(["permission_key" => 'timeout_reply_group.edit']),
+                    Route::delete('/rules/{id:\d+}')->action([RuleController::class, 'destroy'])->defaults(["permission_key" => 'timeout_reply_group.edit']),
                 ),
         ];
     }
